refactor: remove clearCache from lib
